+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.7.0
+
+* Added anonymous mode support - userId is now optional in identifyUser and setUser
+* Added setLocale method for setting user locale independently
+* Added setAnonymousId method for setting custom anonymous IDs
+* Updated Android SDK dependency to 1.6.0
+* Updated iOS SDK dependency to ~> 1.6.1
+
 ## 0.0.1
 
 * Official Flutter wrapper for the Refiner.io Mobile SDK